split bookmarklet.ts into bookmarkletLogic.ts and boomarkletUI.ts
